Move &time to a new Cycle cog for storing Cycle-related commands
