+++ bot.py
@@ -54,7 +54,7 @@ def init():
     pools_db.set_tables()
 
     # Load cogs
-    cogs = ['cogs.image_manipulation', 'cogs.reminders', 'cogs.miscellaneous', 'cogs.loot']
+    cogs = ['cogs.cycle', 'cogs.image_manipulation', 'cogs.reminders', 'cogs.miscellaneous', 'cogs.loot']
 
     for cog in cogs:
         try:
